requirements-build.txt
aggdraw==1.3.19
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
docopt==0.6.2
idna==3.10
imageio==2.37.0
lazy_loader==0.4
networkx==3.4.2
Nuitka==2.6.7
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
psd-tools==1.10.6
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
soupsieve==2.6
tifffile==2025.2.18
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
